+++ system/Validation/Validation.php
@@ -116,12 +116,15 @@ class Validation implements ValidationInterface
      * @param array|null  $data    The array of data to validate.
      * @param string|null $group   The predefined group of rules to apply.
      * @param string|null $dbGroup The database group to use.
+     *
+     * @TODO Type ?string for $dbGroup should be removed.
+     *      See https://github.com/codeigniter4/CodeIgniter4/issues/6723
      */
     public function run(?array $data = null, ?string $group = null, ?string $dbGroup = null): bool
     {
         $data ??= $this->data;
 
-        // i.e. is_unique
+        // `DBGroup` is a reserved name. For is_unique and is_not_unique
         $data['DBGroup'] = $dbGroup;
 
         $this->loadRuleSets();
@@ -184,17 +187,28 @@ class Validation implements ValidationInterface
     }
 
     /**
-     * Runs the validation process, returning true or false
-     * determining whether validation was successful or not.
+     * Runs the validation process, returning true or false determining whether
+     * validation was successful or not.
      *
      * @param array|bool|float|int|object|string|null $value
+     * @param array|string                            $rules
      * @param string[]                                $errors
+     * @param string|null                             $dbGroup The database group to use.
      */
-    public function check($value, string $rule, array $errors = []): bool
+    public function check($value, $rules, array $errors = [], $dbGroup = null): bool
     {
         $this->reset();
 
-        return $this->setRule('check', null, $rule, $errors)->run(['check' => $value]);
+        return $this->setRule(
+            'check',
+            null,
+            $rules,
+            $errors
+        )->run(
+            ['check' => $value],
+            null,
+            $dbGroup
+        );
     }
 
     /**
@@ -204,87 +218,30 @@ class Validation implements ValidationInterface
      * so that we can collect all of the first errors.
      *
      * @param array|string $value
-     * @param array|null   $rules
-     * @param array|null   $data          The array of data to validate, with `DBGroup`.
+     * @param array        $rules
+     * @param array        $data          The array of data to validate, with `DBGroup`.
      * @param string|null  $originalField The original asterisk field name like "foo.*.bar".
      */
     protected function processRules(
         string $field,
         ?string $label,
         $value,
-        $rules = null,
-        ?array $data = null,
+        $rules = null,       // @TODO remove `= null`
+        ?array $data = null, // @TODO remove `= null`
         ?string $originalField = null
     ): bool {
         if ($data === null) {
             throw new InvalidArgumentException('You must supply the parameter: data.');
         }
 
-        if (in_array('if_exist', $rules, true)) {
-            $flattenedData = array_flatten_with_dots($data);
-            $ifExistField  = $field;
-
-            if (strpos($field, '.*') !== false) {
-                // We'll change the dot notation into a PCRE pattern that can be used later
-                $ifExistField   = str_replace('\.\*', '\.(?:[^\.]+)', preg_quote($field, '/'));
-                $dataIsExisting = false;
-                $pattern        = sprintf('/%s/u', $ifExistField);
-
-                foreach (array_keys($flattenedData) as $item) {
-                    if (preg_match($pattern, $item) === 1) {
-                        $dataIsExisting = true;
-                        break;
-                    }
-                }
-            } else {
-                $dataIsExisting = array_key_exists($ifExistField, $flattenedData);
-            }
-
-            unset($ifExistField, $flattenedData);
-
-            if (! $dataIsExisting) {
-                // we return early if `if_exist` is not satisfied. we have nothing to do here.
-                return true;
-            }
-
-            // Otherwise remove the if_exist rule and continue the process
-            $rules = array_filter($rules, static fn ($rule) => $rule instanceof Closure || $rule !== 'if_exist');
+        $rules = $this->processIfExist($field, $rules, $data);
+        if ($rules === true) {
+            return true;
         }
 
-        if (in_array('permit_empty', $rules, true)) {
-            if (
-                ! in_array('required', $rules, true)
-                && (is_array($value) ? $value === [] : trim((string) $value) === '')
-            ) {
-                $passed = true;
-
-                foreach ($rules as $rule) {
-                    if (! $this->isClosure($rule) && preg_match('/(.*?)\[(.*)\]/', $rule, $match)) {
-                        $rule  = $match[1];
-                        $param = $match[2];
-
-                        if (! in_array($rule, ['required_with', 'required_without'], true)) {
-                            continue;
-                        }
-
-                        // Check in our rulesets
-                        foreach ($this->ruleSetInstances as $set) {
-                            if (! method_exists($set, $rule)) {
-                                continue;
-                            }
-
-                            $passed = $passed && $set->{$rule}($value, $param, $data);
-                            break;
-                        }
-                    }
-                }
-
-                if ($passed === true) {
-                    return true;
-                }
-            }
-
-            $rules = array_filter($rules, static fn ($rule) => $rule instanceof Closure || $rule !== 'permit_empty');
+        $rules = $this->processPermitEmpty($value, $rules, $data);
+        if ($rules === true) {
+            return true;
         }
 
         foreach ($rules as $i => $rule) {
@@ -358,6 +315,92 @@ class Validation implements ValidationInterface
         }
 
         return true;
+    }
+
+    /**
+     * @param array $data The array of data to validate, with `DBGroup`.
+     *
+     * @return array|true The modified rules or true if we return early
+     */
+    private function processIfExist(string $field, array $rules, array $data)
+    {
+        if (in_array('if_exist', $rules, true)) {
+            $flattenedData = array_flatten_with_dots($data);
+            $ifExistField  = $field;
+
+            if (strpos($field, '.*') !== false) {
+                // We'll change the dot notation into a PCRE pattern that can be used later
+                $ifExistField   = str_replace('\.\*', '\.(?:[^\.]+)', preg_quote($field, '/'));
+                $dataIsExisting = false;
+                $pattern        = sprintf('/%s/u', $ifExistField);
+
+                foreach (array_keys($flattenedData) as $item) {
+                    if (preg_match($pattern, $item) === 1) {
+                        $dataIsExisting = true;
+                        break;
+                    }
+                }
+            } else {
+                $dataIsExisting = array_key_exists($ifExistField, $flattenedData);
+            }
+
+            if (! $dataIsExisting) {
+                // we return early if `if_exist` is not satisfied. we have nothing to do here.
+                return true;
+            }
+
+            // Otherwise remove the if_exist rule and continue the process
+            $rules = array_filter($rules, static fn ($rule) => $rule instanceof Closure || $rule !== 'if_exist');
+        }
+
+        return $rules;
+    }
+
+    /**
+     * @param array|string $value
+     * @param array        $data  The array of data to validate, with `DBGroup`.
+     *
+     * @return array|true The modified rules or true if we return early
+     */
+    private function processPermitEmpty($value, array $rules, array $data)
+    {
+        if (in_array('permit_empty', $rules, true)) {
+            if (
+                ! in_array('required', $rules, true)
+                && (is_array($value) ? $value === [] : trim((string) $value) === '')
+            ) {
+                $passed = true;
+
+                foreach ($rules as $rule) {
+                    if (! $this->isClosure($rule) && preg_match('/(.*?)\[(.*)\]/', $rule, $match)) {
+                        $rule  = $match[1];
+                        $param = $match[2];
+
+                        if (! in_array($rule, ['required_with', 'required_without'], true)) {
+                            continue;
+                        }
+
+                        // Check in our rulesets
+                        foreach ($this->ruleSetInstances as $set) {
+                            if (! method_exists($set, $rule)) {
+                                continue;
+                            }
+
+                            $passed = $passed && $set->{$rule}($value, $param, $data);
+                            break;
+                        }
+                    }
+                }
+
+                if ($passed === true) {
+                    return true;
+                }
+            }
+
+            $rules = array_filter($rules, static fn ($rule) => $rule instanceof Closure || $rule !== 'permit_empty');
+        }
+
+        return $rules;
     }
 
     /**
@@ -679,6 +722,7 @@ class Validation implements ValidationInterface
 
                     foreach ($placeholderFields as $field) {
                         $validator ??= Services::validation(null, false);
+                        assert($validator instanceof Validation);
 
                         $placeholderRules = $rules[$field]['rules'] ?? null;
 
@@ -699,7 +743,8 @@ class Validation implements ValidationInterface
                         }
 
                         // Validate the placeholder field
-                        if (! $validator->check($data[$field], implode('|', $placeholderRules))) {
+                        $dbGroup = $data['DBGroup'] ?? null;
+                        if (! $validator->check($data[$field], $placeholderRules, [], $dbGroup)) {
                             // if fails, do nothing
                             continue;
                         }