+++ user_guide_src/source/database/configuration/004.php
@@ -1,13 +1,18 @@
 <?php
 
-// MySQLi
-$default = [
-    'DSN' => 'MySQLi://username:password@hostname:3306/database?charset=utf8&DBCollat=utf8_general_ci',
-    // ...
-];
+namespace Config;
 
-// Postgre
-$default = [
-    'DSN' => 'Postgre://username:password@hostname:5432/database?charset=utf8&connect_timeout=5&sslmode=1',
+use CodeIgniter\Database\Config;
+
+class Database extends Config
+{
     // ...
-];
+
+    // MySQLi
+    public array $default = [
+        'DSN' => 'MySQLi://username:password@hostname:3306/database?charset=utf8&DBCollat=utf8_general_ci',
+        // ...
+    ];
+
+    // ...
+}