+++ user_guide_src/source/database/configuration/006.php
@@ -6,7 +6,9 @@ use CodeIgniter\Database\Config;
 
 class Database extends Config
 {
-    public $test = [
+    // ...
+
+    public array $test = [
         'DSN'      => '',
         'hostname' => 'localhost',
         'username' => 'root',