+++ user_guide_src/source/database/query_builder/031.php
@@ -1,7 +1,11 @@
 <?php
 
 // With closure
-$builder->whereIn('id', static fn (BaseBuilder $builder) => $builder->select('job_id')->from('users_jobs')->where('user_id', 3));
+use CodeIgniter\Database\BaseBuilder;
+
+$builder->whereIn('id', static function (BaseBuilder $builder) {
+    $builder->select('job_id')->from('users_jobs')->where('user_id', 3);
+});
 // Produces: WHERE "id" IN (SELECT "job_id" FROM "users_jobs" WHERE "user_id" = 3)
 
 // With builder directly