+++ user_guide_src/source/database/query_builder/058.php
@@ -1,7 +1,11 @@
 <?php
 
 // With closure
-$builder->orHavingNotIn('id', static fn (BaseBuilder $builder) => $builder->select('user_id')->from('users_jobs')->where('group_id', 3));
+use CodeIgniter\Database\BaseBuilder;
+
+$builder->orHavingNotIn('id', static function (BaseBuilder $builder) {
+    $builder->select('user_id')->from('users_jobs')->where('group_id', 3);
+});
 // Produces: OR "id" NOT IN (SELECT "user_id" FROM "users_jobs" WHERE "group_id" = 3)
 
 // With builder directly